smm: Spin on memory location rather than IO port during SMM relocation
During SMM relocation, a scratch I/O port with address PORT_SMI_STATUS is used
to signal completion of SMM relocation. The idea of using this port seems to be
to have a stable location to communicate completion of SMM relocation. Observing
that this location is only used in the non-relocated case, it is possible to use
a memory location instead which makes this signalling more hardware-agnostic.

Having a more hardware-agnostic approach allows for adding VIA chipset support
which doesn't have an equivalent scratch I/O register.

Signed-off-by: Bernhard Beschow <shentey@gmail.com>